+++ app/di.php
@@ -1,6 +1,7 @@
 <?php
 
 $registrar->addInstance(new \GovUKBlogs\WidgetCategoriesDropdownWithSubmit\Register());
+$registrar->addInstance(new \GovUKBlogs\WidgetArchiveDropdownWithSubmit\Register());
 $registrar->addInstance(new \GovUKBlogs\FixRoots());
 $registrar->addInstance(new \GovUKBlogs\OpenGraphImage());
 $registrar->addInstance(new \GovUKBlogs\Embed\YouTube());