+++ CHANGELOG.md
@@ -3,6 +3,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Accessible "Archives with dropdown" widget
+
+### Changed
+
+- Styles on "Categories with dropdown" widget made more consistent with govuk-frontend default styles
+
 ## [3.1.9] - 2020-07-23
 
 ### Removed